Added another log.h; removed redundant trace args
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@16872 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/x11/window.cpp
@@ -1268,7 +1268,7 @@ bool wxAddWindowToTable(Window w, wxWindow *win)
     wxWidgetHashTable->Put((long) w, win);
 
     wxLogTrace("widget", "XWindow 0x%08x <-> window %p (%s)",
-               w, win, win->GetClassInfo()->GetClassName());
+               (unsigned int) w, win, win->GetClassInfo()->GetClassName());
 
     return TRUE;
 }
@@ -1300,7 +1300,7 @@ bool wxAddClientWindowToTable(Window w, wxWindow *win)
     wxClientWidgetHashTable->Put((long) w, win);
 
     wxLogTrace("widget", "XWindow 0x%08x <-> window %p (%s)",
-               w, win, win->GetClassInfo()->GetClassName());
+               (unsigned int) w, win, win->GetClassInfo()->GetClassName());
 
     return TRUE;
 }
